Replace `speed-up-shows` profile w/ script
If it's likely we're watching a show, then the speed will get set to 1.25x as expected. What's different is the behaviour when changing file.

For shows, the last speed set will be used for subsequent files, unless it's being resumed and has its own speed set.

For anything else. the speed will be reset to 1, unless resuming and the file has its own speed set.