subsel: set English SDH subtitle track if present
This will be skipped if the subtitle track automatically selected by mpv belongs to an external file.

Unfortunately, mpv seems to be buggy and won't save selected subtitle tracks to the file's resume file, so if you change track from the selected SDH, go to another file and then go back, SDH will be forcibly selected again by this script.