"Update subpos in `hide_osc()`"
This fixes an issue whereby the sub-pos does not revert to its original
position, when osc-visibility is set to never, WHILE the OSC is still
visible. A more common use case would be to cycle through osc-visibility
using the DEL keybind. Since "always" is before "never", the bug will be
triggered.

The core logic of the subpos code relies on the function osc_visible()
to determine whether the OSC is visible. The main issue stems from the
fact that when osc-visibility is set to "never", osc_visible() won't
get invoked. So, call update_subpos() explicitly in hide_osc().

From dexeonify/mpv-config @ d7adeb42111952ff71be4b3875ac74e05f0885f6